Rename bundle splitting to chunking
